+++ app/src/main/java/com/palone/planahead/screens/taskEdit/ui/taskTypeSelection/typeProperties/cronProperties/RepeatTimeSelector.kt
@@ -26,8 +26,6 @@ fun RepeatTimeSelector(
             onTimeChange(
                 LocalTime.of(hours, minutes)
             )
-        }
-        )
+        })
     }
-
 }
