+++ app/src/main/java/com/palone/planahead/screens/taskEdit/ui/taskTypeSelection/typeProperties/cronProperties/WeekDaysList.kt
@@ -62,6 +62,5 @@ fun Day(
         ) {
             Text(text = field.name[0].toString() + field.name[1].toString().lowercase())
         }
-
     }
 }
