+++ app/src/main/java/com/palone/planahead/ui/theme/Theme.kt
@@ -5,64 +5,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import com.example.compose.md_theme_dark_background
-import com.example.compose.md_theme_dark_error
-import com.example.compose.md_theme_dark_errorContainer
-import com.example.compose.md_theme_dark_inverseOnSurface
-import com.example.compose.md_theme_dark_inversePrimary
-import com.example.compose.md_theme_dark_inverseSurface
-import com.example.compose.md_theme_dark_onBackground
-import com.example.compose.md_theme_dark_onError
-import com.example.compose.md_theme_dark_onErrorContainer
-import com.example.compose.md_theme_dark_onPrimary
-import com.example.compose.md_theme_dark_onPrimaryContainer
-import com.example.compose.md_theme_dark_onSecondary
-import com.example.compose.md_theme_dark_onSecondaryContainer
-import com.example.compose.md_theme_dark_onSurface
-import com.example.compose.md_theme_dark_onSurfaceVariant
-import com.example.compose.md_theme_dark_onTertiary
-import com.example.compose.md_theme_dark_onTertiaryContainer
-import com.example.compose.md_theme_dark_outline
-import com.example.compose.md_theme_dark_outlineVariant
-import com.example.compose.md_theme_dark_primary
-import com.example.compose.md_theme_dark_primaryContainer
-import com.example.compose.md_theme_dark_scrim
-import com.example.compose.md_theme_dark_secondary
-import com.example.compose.md_theme_dark_secondaryContainer
-import com.example.compose.md_theme_dark_surface
-import com.example.compose.md_theme_dark_surfaceTint
-import com.example.compose.md_theme_dark_surfaceVariant
-import com.example.compose.md_theme_dark_tertiary
-import com.example.compose.md_theme_dark_tertiaryContainer
-import com.example.compose.md_theme_light_background
-import com.example.compose.md_theme_light_error
-import com.example.compose.md_theme_light_errorContainer
-import com.example.compose.md_theme_light_inverseOnSurface
-import com.example.compose.md_theme_light_inversePrimary
-import com.example.compose.md_theme_light_inverseSurface
-import com.example.compose.md_theme_light_onBackground
-import com.example.compose.md_theme_light_onError
-import com.example.compose.md_theme_light_onErrorContainer
-import com.example.compose.md_theme_light_onPrimary
-import com.example.compose.md_theme_light_onPrimaryContainer
-import com.example.compose.md_theme_light_onSecondary
-import com.example.compose.md_theme_light_onSecondaryContainer
-import com.example.compose.md_theme_light_onSurface
-import com.example.compose.md_theme_light_onSurfaceVariant
-import com.example.compose.md_theme_light_onTertiary
-import com.example.compose.md_theme_light_onTertiaryContainer
-import com.example.compose.md_theme_light_outline
-import com.example.compose.md_theme_light_outlineVariant
-import com.example.compose.md_theme_light_primary
-import com.example.compose.md_theme_light_primaryContainer
-import com.example.compose.md_theme_light_scrim
-import com.example.compose.md_theme_light_secondary
-import com.example.compose.md_theme_light_secondaryContainer
-import com.example.compose.md_theme_light_surface
-import com.example.compose.md_theme_light_surfaceTint
-import com.example.compose.md_theme_light_surfaceVariant
-import com.example.compose.md_theme_light_tertiary
-import com.example.compose.md_theme_light_tertiaryContainer
 
 
 private val LightColors = lightColorScheme(
